+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.3] Latest emoji data
+
+- 1.0.3 Update import script and get latest emoji data.
+
 ## [1.0.2] Bug fixes
 
 - 1.0.2 Fix for skin tone modifiers to show woman emoji. Only hides emoji with more than one skin tone modifier.